Factor out `BoardServerProvider`.
