+++ packages/board-server/src/server/boards/invoke.ts
@@ -4,137 +4,24 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { getDataStore } from "@breadboard-ai/data-store";
+import AgentKit from "@google-labs/agent-kit/agent.kit.json" with { type: "json" };
 import {
   asRuntimeKit,
   createLoader,
   inflateData,
-  type ChangeNotificationCallback,
-  type GraphDescriptor,
-  type GraphProvider,
-  type GraphProviderCapabilities,
-  type GraphProviderExtendedCapabilities,
-  type GraphProviderStore,
   type KitManifest,
 } from "@google-labs/breadboard";
-import type { ApiHandler, BoardParseResult } from "../types.js";
 import { run, type HarnessRunResult } from "@google-labs/breadboard/harness";
 import { fromManifest } from "@google-labs/breadboard/kits";
-import AgentKit from "@google-labs/agent-kit/agent.kit.json" with { type: "json" };
-import GeminiKit from "@google-labs/gemini-kit";
-import TemplateKit from "@google-labs/template-kit";
-import JSONKit from "@google-labs/json-kit";
 import Core from "@google-labs/core-kit";
-import { asInfo, getStore } from "../store.js";
+import GeminiKit from "@google-labs/gemini-kit";
+import JSONKit from "@google-labs/json-kit";
+import TemplateKit from "@google-labs/template-kit";
 import { secretsKit } from "../proxy/secrets.js";
-import { getDataStore } from "@breadboard-ai/data-store";
-
-class BoardServerProvider implements GraphProvider {
-  #path: string;
-
-  name = "Board Server Provider";
-
-  constructor(path: string) {
-    this.#path = path;
-  }
-
-  async ready(): Promise<void> {
-    return;
-  }
-
-  isSupported(): boolean {
-    return true;
-  }
-
-  canProvide(url: URL): false | GraphProviderCapabilities {
-    return url.href.endsWith(this.#path)
-      ? {
-          load: true,
-          save: false,
-          delete: false,
-        }
-      : false;
-  }
-
-  extendedCapabilities(): GraphProviderExtendedCapabilities {
-    return {
-      modify: false,
-      connect: false,
-      disconnect: false,
-      refresh: false,
-      watch: false,
-    };
-  }
-
-  async load(url: URL): Promise<GraphDescriptor | null> {
-    const store = getStore();
-    const { userStore, boardName } = asInfo(this.#path);
-    if (!userStore || !boardName) {
-      return null;
-    }
-    const graph = JSON.parse(await store.get(userStore, boardName));
-    return graph as GraphDescriptor;
-  }
-
-  async save(
-    url: URL,
-    graph: GraphDescriptor
-  ): Promise<{ result: boolean; error?: string }> {
-    throw new Error("Method not supported.");
-  }
-
-  createBlank(url: URL): Promise<{ result: boolean; error?: string }> {
-    throw new Error("Method not supported.");
-  }
-
-  create(
-    url: URL,
-    graph: GraphDescriptor
-  ): Promise<{ result: boolean; error?: string }> {
-    throw new Error("Method not supported.");
-  }
-
-  delete(url: URL): Promise<{ result: boolean; error?: string }> {
-    throw new Error("Method not supported.");
-  }
-
-  connect: (location?: string, auth?: unknown) => Promise<boolean> =
-    async () => {
-      throw new Error("Method not supported.");
-    };
-
-  disconnect: (location: string) => Promise<boolean> = async () => {
-    throw new Error("Method not supported.");
-  };
-
-  refresh: (location: string) => Promise<boolean> = async () => {
-    throw new Error("Method not supported.");
-  };
-
-  createURL: (location: string, fileName: string) => Promise<string | null> =
-    async () => {
-      throw new Error("Method not supported.");
-    };
-
-  watch: (callback: ChangeNotificationCallback) => void = async () => {
-    throw new Error("Method not supported.");
-  };
-
-  parseURL(url: URL): { location: string; fileName: string } {
-    throw new Error("Method not supported.");
-  }
-
-  restore: () => Promise<void> = async () => {
-    throw new Error("Method not supported.");
-  };
-
-  items: () => Map<string, GraphProviderStore> = () => {
-    throw new Error("Method not supported.");
-  };
-
-  startingURL: () => URL | null = () => {
-    throw new Error("Method not supported.");
-  };
-}
+import { getStore } from "../store.js";
+import type { ApiHandler, BoardParseResult } from "../types.js";
+import { BoardServerProvider } from "./utils/board-server-provider.js";
 
 const kits = [
   secretsKit,
